Moved OpenAPI /generated folder to inside of src/app as src/app/generated-data-api
