+++ src/app/data.service.ts
@@ -3,10 +3,10 @@ import {Demographic} from './datamodel/demographics';
 import {SubjectDataService} from './subject-data-service.service';
 import {CareplanService} from './careplan.service';
 // import {ConditionLists} from './datamodel/conditionLists';
-import {MccCarePlan} from '../../generated/models/MccCarePlan';
+import {MccCarePlan} from './generated-data-api';
 import {SocialConcerns} from './datamodel/socialconcerns';
 // import {ConditionLists} from './datamodel/conditionLists';
-import {ConditionLists} from '../../generated';
+import {ConditionLists} from './generated-data-api';
 import {TargetValue} from './datamodel/targetvalue';
 import {
   dummyPatientId,
@@ -23,7 +23,7 @@ import {
   mockGoalList,
   mockMedicationSummary
 } from './datamodel/mockData';
-import {GoalLists} from '../../generated';
+import {GoalLists} from './generated-data-api';
 import {MedicationSummary} from './datamodel/medicationSummary';
 import {Education} from './datamodel/education';
 import {Referral} from './datamodel/referral';
