+++ src/app/datamodel/mockData.ts
@@ -1,14 +1,14 @@
 import {TargetValue} from './targetvalue';
-// import {GoalLists} from '../../../generated/models/GoalLists';
+// import {GoalLists} from '../../../generated-data-api/models/GoalLists';
 import {GoalLists} from './goallists';
 import {MedicationSummary} from './medicationSummary';
 import {Education} from './education';
 import {Referral} from './referral';
 import {Contact} from './contact';
 import {Demographic} from './demographics';
-import {ConditionLists} from '../../../generated/models/ConditionLists';
+import {ConditionLists} from '../generated-data-api';
 import {Concern} from './socialconcerns';
-import {MccCarePlan} from '../../../generated/models/MccCarePlan';
+import {MccCarePlan} from '../generated-data-api';
 
 export const mockContacts: Contact[] = [
   {
