+++ src/app/generated-data-api/core/OpenAPI.ts
