+++ src/app/generated-data-api/core/Result.ts
