+++ src/app/generated-data-api/core/getQueryString.ts
