+++ src/app/generated-data-api/core/isSuccess.ts
