+++ src/app/generated-data-api/core/requestUsingFetch.ts
