+++ src/app/generated-data-api/models/Bounds.ts
