+++ src/app/generated-data-api/models/ConditionSummary.ts
