+++ src/app/generated-data-api/models/Dictionary.ts
