+++ src/app/generated-data-api/models/GenericType.ts
