+++ src/app/generated-data-api/models/GoalSummary.ts
