+++ src/app/generated-data-api/models/MccCondition.ts
