+++ src/app/generated-data-api/models/MccDate.ts
