+++ src/app/generated-data-api/models/MccDuration.ts
