+++ src/app/generated-data-api/models/MccGoal.ts
