+++ src/app/generated-data-api/models/MccIdentifer.ts
