+++ src/app/generated-data-api/models/MccInstant.ts
