+++ src/app/generated-data-api/models/MccObservation.ts
