+++ src/app/generated-data-api/models/MccPatient.ts
