+++ src/app/generated-data-api/models/MccPeriod.ts
