+++ src/app/generated-data-api/models/MccQuantity.ts
