+++ src/app/generated-data-api/models/MccRange.ts
