+++ src/app/generated-data-api/models/MccRatio.ts
