+++ src/app/generated-data-api/models/MccReference.ts
