+++ src/app/generated-data-api/models/MccSampledData.ts
