+++ src/app/generated-data-api/models/MccSimpleQuantity.ts
