+++ src/app/generated-data-api/models/MccTime.ts
