+++ src/app/generated-data-api/models/ObservationComponent.ts
