+++ src/app/generated-data-api/models/ReferenceRange.ts
