+++ src/app/generated-data-api/services/ConditionControllerService.ts
