+++ src/app/subject-data-service.service.ts
@@ -7,7 +7,7 @@ import { Demographic } from './datamodel/demographics';
 import { MessageService } from './message.service';
 import {Concern, SocialConcerns} from './datamodel/socialconcerns';
 import {environment} from '../environments/environment';
-import {ConditionLists} from '../../generated';
+import {ConditionLists} from './generated-data-api';
 
 @Injectable({providedIn: 'root'})
 export class SubjectDataService{
